feat: profile avatar use case for widgetbook

+++ apps/widgetbook/lib/use_case/profile/profile_button_use_case.dart
@@ -43,7 +43,7 @@ Widget buildProfileButton(BuildContext context) {
 				BlocProvider<ProfileCubit>.value(value: profileCubit),
 			],
 			child: ProfileButton(
-        size: size,
+        size: size.toDouble(),
         profileImageProvider: AssetImage('assets/images/avatar.jpg'),
       ),
 		),
